+++ EssentialFeediOS/FeedUI/Composers/WeakRefVirtualProxy.swift
@@ -1,0 +1,29 @@
+//
+//  WeakRefVirtualProxy.swift
+//  EssentialFeediOS
+//
+//  Created by Jair Moreno Gaspar on 05/01/21.
+//  Copyright © 2021 Jair Moreno G. All rights reserved.
+//
+
+import UIKit
+
+final class WeakRefVirtualProxy<T: AnyObject> {
+    private weak var object: T?
+    
+    init(_ object: T) {
+        self.object = object
+    }
+}
+
+extension WeakRefVirtualProxy: FeedLoadingView where T: FeedLoadingView {
+    func display(_ viewModel: FeedLoadingViewModel) {
+        object?.display(viewModel)
+    }
+}
+
+extension WeakRefVirtualProxy: FeedImageView where T: FeedImageView, T.Image == UIImage {
+    func display(_ model: FeedImageViewModel<UIImage>) {
+        object?.display(model)
+    }
+}
